Operator adapt to short type

Created-by: liushuai 30024241
Author-id: 520315
MR-id: 12412488
Commit-by: l30024241
Merged-by: lizheng 00465000
E2E-issues: AR20220730929362;SR20220730929292
Description:
operator adapt to short type

See merge request Kunpeng-Computing-DC/BigData/OmniOperatorJIT!236

+++ core/src/operator/window/window.cpp
@@ -284,7 +284,7 @@ void WindowOperator::InitResultVectors(const std::vector<DataTypePtr> &outputTyp
                 break;
             }
             case OMNI_SHORT: {
-                vecBatchField->SetVector(colIndex, new IntVector(vecAllocator, rowCountField));
+                vecBatchField->SetVector(colIndex, new ShortVector(vecAllocator, rowCountField));
                 break;
             }
             case OMNI_VARCHAR:
